+++ Changelog.md
@@ -1,3 +1,8 @@
+RELEASE_06
+- Added "bingo" intent: Vector pulls out the numbers from 1 to 90. 
+  To pull a number, touch Vector. You can also shake it or just caress it, it reacts to touch on
+  tha back. Press the back button to quit. Useful for bingo nights...
+
 RELEASE_05
 - Bugfix: add localized response of the "roll a die" intent
 - Bugfix: fix localization for "your name is" intent